Move diff-tests to own file
